Modify title and Home button and add logo details
Changed the title element from "Document" to name of the restaurant.
Added some styling to the Home button to test out the CSS loading from
Webpack. Added the restaurant logo to the top of the page along with a
comment consisting of the credits for the logo to be used later in the
footer.