Add images for Gallery and About to project files
Added menu food images and another restaurant image to the project files
for the Gallery and About sections on the homepage. Added credits for
the images as comments at the bottom of the contents section for the
footer later on. Added restaurant image to About section to be used as
a low opacity/"faded" background image behind the text later on.