Refactor and edit JS code for image debugging
Refactored and modified code inside of JS in an attempt to debug the
404 error "broken images" that are on the homepage. Error still
persists. Added comments to webpack config file regarding possibly
turning watch mode "on" (though it should already BE on) and a note to
find out why the images assets aren't listed as a dependency in the
package.json file and whether or not they should be (would this solve
the 404 error?).