Comment out all div content elements in HTML
"Removed" all elements under the <div class="content"> per assignment
instructions. By "removed", I mean commented out so I could use it as
a reference in the future for the project (may still need the code).
Added a comment regarding the broken image "error" at the top of the
JS file regarding why the "error" is showing. This may solve itself in
the future as we go along (Webpack only deals with JS right now).