Uncomment content div in HTML & create new JS file
Uncomment the content div itself in the HTML as that's what was actually
stated in the project instructions (content div was commented out along
with it's children which was incorrect). Now, ONLY the children of the
content div are commented out for reference. Created a new JS file for
all the homepage elements to be created dynamically under via DOM.