Append homepage elements to the content div
Refactored both the index and home JS files so the content div now lives
on the index.js file as I couldn't find another way to append the
homepage elements by just leaving the content DOM on the home.js file.
Homepage function is now appended to the content div via index.js.
Added a dev comment at the bottom of the home.js file regarding a
potential concern with how the homepage sections are added to the
content div (the main sections). Will test to see if everything loads
correctly as is.