Fix webpack not recognizing logo or homepage files
Resolved the terminal error where the logo image and the home.js file
were not being recognized. Configuration was correct but it turned out
to be a few typos in the index.js file. Fixed other typos on the img
files linked inside the homepage function under home.js as well as the
CSS file.