Fix missing text blurb under About Us section
Appended the text content under the About Us section of the homepage as
it wasn't showing at all (was never appended to begin with). Refactored
some comments regarding a possible forEach method for the Gallery
section. Added some background image styling to the Our Philosophy
image but the image isn't loading correctly. Nor are the other images
on the homepage except for the logo (which is set as a background image)