Modify some images from relative to absolute paths
Changed some of the targeted paths for the images from a relative to an
absolute path, indicating that something is wrong with the path that's
targeting the same saved images in the project. Added a comment in the
index.js to indicate instructions on how to implement the menu buttons.