Modify image file paths from relative to absolute
Updated logo file path to absolute in the HTML. Updated all file paths
for all images to be absolute instead of relative in the home.js file
with the exception of one (on purpose). Will experiment with inspecting
the 404 error in DevTools and researching CORS for the final image to
see if I can come up with another solution for the broken images error
due to the relative filepaths. Added another comment in home.js
regarding next steps on the forEach loop in the gallery subtext section.
Commented out imported logo and related code in the index.js file due
to logo image being updated in the HTML (see above). Commented out logo
styling for the time being in the CSS, though did update it's file path
to absolute just in case. Updated the file path for the image in the
Philosophy section in the CSS as well.