Modify final image filepath and dev comments
Changed the last image filepath from relative to absolute after some
research into the errors being generated in DevTools. Absolute filepath
seems to be better for this type of project (relative is more JS but
also more lines of code that seem unnecessary). Added some TypeError
comments to look into in the webpack config and the index.js