Modify Testimonial section and remove comments
Centered the Testimonial section of the homepage by centering all the
text with correct positioning between the quote and the author sections.
Attempted to center the About Us section the same way but was not
successful. Left a comment with a possible reason why. Deleted comments
in the Testimonial section of home.js as they are no longer needed on
the "TODO" list since everything is styled correctly in that section.