Modify styling of header elements on home page
Centered all the elements in the header section of the home page.
Resized the logo to make it more proportionate to the rest of the page
and make it look more like an actual logo. Gave the navbar some simple
padding for now, but will be styled further. Added more comments in
the CSS on next steps for styling in multiple sections.