Resize media section of homepage
Resized the media image on the first section of the homepage to take up
the full width. Will still need to adjust sizing as there is now a
scrollbar on the bottom. Refactored background image in the Our
Philosophy section although more styling for the entire section is still
needed.