Style the Our Philosophy section & modify colors
Modified the styling and positioning of all elements in the Our
Philosophy section of the homepage. Modified the background and text
colors to reflect the branding color of the site a little better (this
still needs work though). Commented out all JS that had to do with
appending the background image of the Our Philosophy section (may delete
this entirely later).