Modify header section to give it fixed positioning
Modified the styling of the header section so that the header stays
fixed to the page as the user scrolls up and down. Shrunk the navbar
section as the header was taking up too much space on scroll. May
experiment with ONLY giving the navbar fixed positioning so more of
the rest of the page can be seen and restore navbar to original size.
Copied comments regarding h2 element in About section over to the Our
Philosophy section as well. Removed empty line of code in HTML that
wasn't necessary.