Modify navbar and nav button styling
Revised the button styling inside the navbar. The new look makes the
header slightly smaller and take up a bit less space on the page. The
buttons look slightly more modern with the new "pill" shape. Gave the
Home button an "active" status in the HTML which will be toggled to
the button the user clicks on future JS updates.