Add menu items with description, image and price
Added several elements to the menu page including the menu header and a
menu section that currently will contain three menu items. Each item
has had an image, description and price added to it along with a Buy
button for each. Currently the buttons have additional copy code but
will be changed to a forEach method to reduce code.