Modify item names and descriptions on menu page
Removed all the placeholder descriptions for all the menu items, which
was a short description of the first item, and modified the first menu
item description with a better, more tantalizing one. Did the same for
the other two menu items. Removed the placeholder names for the second
and third items and replaced them with their actual names. After some
thinking, decided to bring back the original buttons that were commented
out instead of putting them in a forEach loop. The goal is to have an
MVP and leaving the original button setup the way it is will be quicker
for now. The plan will be to refactor the menu page code with some
forEach methods but only after the MVP is up.