Modify nav button click functionality and styling
Changed the nav button click functionality in the index.js file from an
'if/else' statement to a 'switch' statement for testing purposes. Will
keep if it works in a future update. Changed the targeting syntax of the
nav button in the CSS from targeting the 'id' to targeting whatever
button has 'aria-current' on it at the time of interaction. Removed
TODO comments that were no longer needed and added some new ones in all
the JS files.