Modify button selectors, forEach & menu JS code
The forEach method for the nav buttons wasn't working properly, causing
the entire homepage as well as the footer to not display at all (only
the logo, header and navbar were showing). The method was not targeting
the buttons properly (the nav element itself was being targeted) for
the forEach method. Refactored the nav buttons in the HTML first to give
them all a class of the same name and have their previous class names be
their unique id tags instead (hopefully aria-current will work later).
This allowed me to target the buttons in the DOM correctly and refactor
the forEach method in index.js file. Homepage and footer now show again
correctly, but the buttons themselves still don't work. Added a line in
menu.js file that appended the menu section to the content div. This is
similar to how the homepage sections were appended in home.js but I will
test this later by removing them in their respective JS code as I don't
think they're needed to be honest.