Fix issue of navbar buttons not working on click
Was able to resolve the issue of the navbar buttons not responding to
any click events by modifying the switch statement cases to look for
the text content in the buttons and not some string in the button
object itself. The tabbed browsing still doesn't work though due to a
separate Uncaught TypeError. Will debug this next.