Modify position behavior of header and menu page
Changed the position behavior of the entire header to be sticky with a
top position of 0px to stick right away. With position being fixed, the
menu header was hidden behind the header because of the way fixed
positioning works. Viewport width for the images on the homepage is now
off again so will fix in another update. Menu elements are mostly
positioned correctly. Uncaught TypeError involving 'Node' still showing
but the navigation works aside from that.