Modify button DOM manipulation for menu page
Tried to modify some of the JS for the Add to Order buttons on the menu
page in an attempt to resolve the issue where the button is showing up
on the 2nd item but not the 1st or 3rd. Was unsuccessful. Seems like a
CSS issue with the grid but not sure.