Resolve issue of Order buttons not showing
Was able to resolve the issue of the Order button showing up on one item
but not the other items on the menu page. Ended up giving each item
it's own unique button with unique naming convention after failing to
put together a forEach method that would add the same type of button to
all the menu items.