Revise index.js code to debug TypeError
Revised the index.js log along with changing the naming conventions in
both the homepage and menupage JS files to fix the TypeError parameter
bug showing up every time a user switches pages on the nav bar. Have
not been successful so far. Will try to track it down in DevTools
instead.