Modify code in an attempt to fix TypeError
Tried to add a 'try catch' method to further drill down on the TypeError
I was getting on the last line of code (parameter not being a Node when
appendChild tries to execute). Also moved the console.log above which
allowed the console to correctly identify the typeof, which is a
function. Attempted to add a parameter to the function since it was
empty, but didn't work. I've now realized it's the function ITSELF that
isn't a Node when appendChild is called. Possibly changing the setup of
how the content div is being called may resolve the issue.