Resolve TypeError parameter node error message
Made the necessary changes to all JS files representing each page for
the website. The issue was that none of the functions that generate the
content for each page was actually returning anything. Added another
container to the homepage function to put all the sections into and
returned that container to be added to div#content. Container already
existed in menupage so it was simply a matter of returning it. Error no
longer shows up. Bars are now separating each section of the homepage,
so made a note in CSS to address this in the future.