Add the message section to the Contact page
Added the message section and included all necessary elements to the
Contact form on the Contact tab of the webpage. Also added some test
'placeholder' text to the textarea to see if it works (may need it to
be 'textContent' or not needed at all). Changed the email section's
type and added a placeholder email as well. Exported and imported the
contact page function and added the necessary switch case on the main
index file to load the page on button click.