Modify styling of Contact page along with buttons
Added styling to the Contact page similar to the Menu page as a base,
then styled the form itself. Gave the 'Add to Order' buttons on the Menu
page and the 'Submit' button on the Contact form a hover effect. Will
test later to see if it's a color that works.