Revise color of other elements to match logo
Changed the color of several elements that were a standard blue to the
hue that matches the background color of the business logo. These
included all buttons and any background colors behind text. Changed the
hover effect color to closely match the new color as well. Revised
some comments from TODO (before the project finishes) to BRANCH (to-do
AFTER the project MVP is pushed to GitHub).