Add sliding page animation for testing purposes
Added a sliding CSS animation, including keyframes and placing the
'display: none' setting where needed. Added a slide animation function
to the index JS, then added that function to the click event so the
animation knows what to do depending on the button that was clicked.
Added a preventDefault to the contact form page as a precaution.