Add HtmlWebpackPlugin setup & modify comments
Set up HtmlWebpackPlugin in the webpack config file. Revised filename
in the output as well as added more notes to revert to previous version
of output if needed. Re-enabled watchFiles under the devServer per TOP
instructions on how to handle the template HTML file when modified.
Updated formatting of description in package.json file along with file
automatically updating the dev dependencies with html-webpack-plugin.
Commented out both the stylesheet tag in the head of the original HTML
document as well as the script tag at the bottom of the body. Set up
HtmlWebpackPlugin to automatically generate both of these in the same
spots that they were manually added. A new bundle JS file and a new
HTML file were created by html-webpack-plugin automatically.