Modify background colors and text
Modified some of the background colors and text color on the main page
to make the website easier on the user's eyes for more devices. Added
a branch note in the home.js module to add a container specifically for
the gallery section so we can change the background color on that entire
section in the future.