Add container to entire Gallery section
Added a container for the Gallery section to place all gallery elements
under. Previously the gallery header and the image section were in
their own separate spaces with nothing to contain both. This made
changing the background color difficult. Added a background color to
the new container to try and resolve that issue. Will test in a bit
before working on changing the image section to a carousel.