Attempt to refactor code to include carousel
Attempted to refactor the code for the gallery section of the homepage
module to turn it into a 3D image carousel. The attempt failed.
Commented out all the code changes and reverted back to previous version
of the web page, but left the 5th image of the gallery in for now.

Removed the (e) within the nav button logic after testing the buttons
without it. Everything still functions fine so it doesn't seem like its
needed.