Refactor image carousel code to fix bugs & styling
Refactored all the code in the Gallery section of the homepage function
in the home module. Needed to end up adding another div wrapper that
the "gallery frame" wrapper that contained the images would be inside
of. Reduced the number of carousel dots in the for loop due to two
images being the same and removed all code that generated the duplicate
image. Made all appropriate element appending changes that were needed.

Reorganized the button logic for all carousel navigation buttons (prev,
next and dots) as well as the function that sets the active slide to
the bottom of the index module. The key was that it needed to be under
the code that appended the homepage elements to the UI since none of
the logic worked before due to the logic loading before the actual
UI elements.

Restyled the image carousel in the CSS to condense all buttons and the
gallery frame itself closer together. Was able to "stack" the images
on top of each other via CSS Grid as opposed to before when they were
still showing as separate images across the width of the page so the
carousel would not display the images correctly.

Image carousel is now complete!